+++ packages/blockera/php/Routes/api.php
@@ -30,9 +30,10 @@ try {
         $routes->post('telemetry/opt-in', [ Blockera\Telemetry\Http\Controllers\OptInController::class, 'optIn' ]);
     }
 
-    if ( blockera_telemetry_bug_detector_logger_is_off()) {
+    if (blockera_telemetry_bug_detector_logger_is_off()) {
 
-        $routes->post('telemetry/log-error', [ Blockera\Telemetry\Http\Controllers\DebugLoggerController::class, 'log' ]);
+        $routes->post('telemetry/log-error', [ Blockera\Telemetry\Http\Controllers\BugDetectorAndReporterController::class, 'log' ]);
+        $routes->post('telemetry/log-error/status', [ Blockera\Telemetry\Http\Controllers\BugDetectorAndReporterController::class, 'status' ]);
     }
 
 } catch (Exception $exception) {